+++ flickit-advice-engine/src/main/java/org/flickit/assessment/advice/application/port/in/advicenarration/GetAdviceNarrationUseCase.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.advice.application.port.in;
+package org.flickit.assessment.advice.application.port.in.advicenarration;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.EqualsAndHashCode;
