+++ flickit-advice-engine/src/test/java/org/flickit/assessment/advice/application/service/advicenarration/GetAdviceNarrationServiceTest.java
@@ -1,8 +1,8 @@
-package org.flickit.assessment.advice.application.service.advice;
+package org.flickit.assessment.advice.application.service.advicenarration;
 
 import org.flickit.assessment.advice.application.domain.AdviceNarration;
 import org.flickit.assessment.advice.application.domain.AssessmentResult;
-import org.flickit.assessment.advice.application.port.in.GetAdviceNarrationUseCase;
+import org.flickit.assessment.advice.application.port.in.advicenarration.GetAdviceNarrationUseCase;
 import org.flickit.assessment.advice.application.port.out.advicenarration.LoadAdviceNarrationPort;
 import org.flickit.assessment.advice.application.port.out.assessmentresult.LoadAssessmentResultPort;
 import org.flickit.assessment.common.application.MessageBundle;
